% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideo.R
\name{ideo}
\alias{ideo}
\title{Create an ideogram}
\usage{
ideo(chroms = NULL)
}
\arguments{
\item{chroms}{A data.fram describing the chromosomes.}
}
\value{
An invisible NULL.
}
\description{
Create an ideogram
}
\examples{
ideo()

}
