% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGM.R
\name{plotGM}
\alias{plotGM}
\title{Plot a genetic map}
\usage{
plotGM(map)
}
\arguments{
\item{map}{An object of class "cross" from the package 'qtl' or a list of vectors.}
}
\value{
A ggplot object.
}
\description{
Plot a genetic map
}
\examples{

fake_gm <- structure(
  list(
    `1` = structure(c(D1M430 = 0, D1M318 = 9.8, D1M212 = 13.11, D1M437 = 37.11,
                      D1M490 = 53.51, D1M15 = 79.71, D1M17 = 103.81), class = "A"),
    `2` = structure(c(D2M359 = 0, D2M241 = 21.8, D2M336 = 44.8, D2M398 = 50.2,
                      D2M51 = 72.1), class = "A"),
    `3` = structure(c(D3M227 = 0, D3M10 = 19.7, D3M17 = 32.8), class = "A"),
    `4` = structure(c(D4M286 = 0, D4M145 = 20.7, D4M12 = 36, D4M310 = 52.4), class = "A")),
  class = "map")
p <- plotGM(fake_gm)
p
p <- p + ggtitle("Best linkage map ever")
p
p <- p + annotate(geom="text", x=2.4, y=25, label="Super\nmarker",
                  color="red", size = 3)
p
p + annotate(geom="rect", xmin = 1.5, xmax = 2.4, ymin = 40, ymax = 60,
             color="red", size = 1, fill = NA)






}
