% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Read BLAST custom tabular data}
\usage{
read_blast(file)
}
\arguments{
\item{file}{A file in tabular format.}
}
\value{
A data.frame.
}
\description{
Read BLAST custom tabular data
}
\examples{
# read_blast()

}
