% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stockholm.R
\name{read_stockholm}
\alias{read_stockholm}
\title{Read Stockholm format data}
\usage{
read_stockholm(file)
}
\arguments{
\item{file}{A file in Stockholm format.

\url{https://en.wikipedia.org/wiki/Stockholm_format}}
}
\value{
A list.
}
\description{
Read Stockholm format data
}
\examples{
# read_stockholm()

}
