% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR_to_qtl2.R
\name{vcfR_to_qtl2}
\alias{vcfR_to_qtl2}
\title{Convert vcfR to qtl2 format}
\usage{
vcfR_to_qtl2(
  vcf,
  prefix = "MYDATA",
  geno_file = TRUE,
  pmap = TRUE,
  transpose = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vcf}{A vcfR object}

\item{prefix}{A prefix for file names}

\item{geno_file}{Create a geno_file for qtl2}

\item{pmap}{Create a pmap file (physical map)}

\item{transpose}{transpose the VCF data}

\item{verbose}{report verbose messages}
}
\value{
An invisible NULL.
}
\description{
Convert vcfR to qtl2 format
}
\examples{
# vcfR_to_qtl2()

}
